Conflicts:
	cheatingday/src/test/java/com/icia/cheatingday/ManagerDaoTest.java
	cheatingday/src/test/java/com/icia/cheatingday/ManagerServiceTest.java
	cheatingday/target/classes/com/icia/cheatingday/freeboard/dto/FreeBoardDto$DtoForRead$DtoForReadBuilder.class
	cheatingday/target/classes/com/icia/cheatingday/freeboard/dto/FreeBoardDto$DtoForRead.class
	cheatingday/target/classes/com/icia/cheatingday/freeboard/dto/FreeBoardDto$DtoForUpdate.class
	cheatingday/target/classes/com/icia/cheatingday/freeboard/dto/FreeBoardDto$DtoForWrite.class
	cheatingday/target/classes/com/icia/cheatingday/freeboard/service/FreeBoardService.class
	cheatingday/target/m2e-wtp/web-resources/META-INF/maven/com.icia/cheatingday/pom.properties
	cheatingday/target/test-classes/com/icia/cheatingday/ManagerServiceTest.class

+++ cheatingday/src/main/java/com/icia/cheatingday/admin/service/rest/AdminRestService.java
@@ -4,24 +4,27 @@ import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
 
 import com.icia.cheatingday.admin.dao.*;
-import com.icia.cheatingday.manager.entity.*;
-import com.icia.cheatingday.user.dao.*;
+import com.icia.cheatingday.exception.*;
+import com.icia.cheatingday.review.dao.*;
 import com.icia.cheatingday.user.entity.*;
 
 @Service
 public class AdminRestService {
 
-	//@Autowired
-	//private ReviewDao rdao;
+	@Autowired
+	private ReviewDao rdao;
 	@Autowired
 	private AdminDao adao;
 
-/*	public void deleteReport(int rNo) {
+	public void deleteReport(int rNo) {
 		Review review = rdao.findById(rNo);
+		if(review==null)
+			throw new JobFailException("해당 리뷰를 찾을 수 없습니다");
 		rdao.delete(rNo);
-	}*/
-	public int enabledM(ManagerEntity manager) {
-		return adao.enabledM(manager);
+		
+	}
+	public int enabledM(int mNum) {
+		return adao.enabledM(mNum);
 	}
 	
 }
