+++ cheatingday/src/main/java/com/icia/cheatingday/center/service/rest/QnARestService.java
@@ -1,17 +1,22 @@
 package com.icia.cheatingday.center.service.rest;
 
+import java.io.*;
 import java.time.*;
 import java.time.format.*;
 import java.util.*;
+import java.util.regex.*;
 
 import org.modelmapper.*;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
+import org.springframework.util.*;
+import org.springframework.web.multipart.*;
 
-import com.icia.cheatingday.admin.dao.*;
+import com.fasterxml.jackson.databind.*;
 import com.icia.cheatingday.center.dao.*;
 import com.icia.cheatingday.center.dto.*;
 import com.icia.cheatingday.center.entity.*;
+import com.icia.cheatingday.exception.*;
 import com.icia.cheatingday.manager.dao.*;
 
 @Service
@@ -25,32 +30,42 @@ public class QnARestService {
 	@Autowired
 	private ManagerDao mdao;
 	@Autowired
-	private AdminDao adao;
-	@Autowired
 	private ModelMapper mapper;
+	@Value("${imageFolder}")
+	private String imageFolder;
+	@Value("${imagePath}")
+	private String imagePath;
+	@Autowired
+	private ObjectMapper objectmapper;
+	
+	Pattern ckImagePattern = Pattern.compile("src=\".+\"\\s");
 	
 	public QnADto.DtoForRead read(Integer qNo, String username){
 		QnA qna = qdao.findById(qNo);
+		if(qna==null)
+			throw new JobFailException("해당 문의를 찾을수 없습니다");
 		QnADto.DtoForRead dto = mapper.map(qna,QnADto.DtoForRead.class);
 		String str = qna.getQWriteTime().format(DateTimeFormatter.ofPattern("yyyy년 MM월 dd일"));
+		dto.setMUsername(mdao.findMusernameByMnum(dto.getMNum()));
 		dto.setQWriteTimeStr(str);
-		dto.setMIrum(mdao.findById(dto.getMUsername()).getMIrum());
+		dto.setMIrum(mdao.findMirumeByMnum(dto.getMNum()));
 		dto.setQCategory(qcdao.findById(dto.getQCano()));
-		if(qna.getQIscomment()==true)
-			dto.setComments(qndao.findAllByQno(dto.getQNo()));
+		dto.setComments(qndao.findAllByQno(dto.getQNo()));
 		return dto;
 	}
-	public List<QnAComment> writeQComment(QnAComment qnAComment){
+	
+	public List<QnAComment> writeQComment(QnAComment qnAComment, String ausername){
 		qnAComment.setQcWriteTime(LocalDateTime.now());
-		String commentStr = qnAComment.getQcContent().replaceAll("(\r\n|\r|\n|\n\r)", "<br>");
-		qnAComment.setQcContent(commentStr);
-		qnAComment.setAUsername(adao.findById(qnAComment.getAUsername()));
+		String comment = qnAComment.getQcContent().replaceAll("\n", " ");
+		qnAComment.setQcContent(comment);
 		qndao.insert(qnAComment);
 		qdao.update(QnA.builder().qNo(qnAComment.getQNo()).qIscomment(true).build());
 		return qndao.findAllByQno(qnAComment.getQNo());
 	}
 	public List<QnAComment> deleteComment(Integer qNo, Integer qcNo, String username) {
 		QnAComment qnAComment = qndao.findById(qcNo);
+		if(qnAComment.getAUsername().equals(username)==false)
+			throw new JobFailException("댓글을 찾을 수 없습니다");
 		qndao.delete(qcNo);
 		qdao.update(QnA.builder().qNo(qnAComment.getQNo()).qIscomment(false).build());
 		return qndao.findAllByQno(qNo);
@@ -58,7 +73,11 @@ public class QnARestService {
 	
 	public void deletQna(Integer qNo, String username) {
 		QnA qnA = qdao.findById(qNo);
-		/*String content = qnA.getQContent();
+		if(qnA==null)
+			throw new JobFailException("헤당 문의를 찾을 수 없습니다");	
+		if(mdao.findMusernameByMnum(qnA.getMNum()).equals(username)==false)
+			throw new UserNotFoundException();
+		String content = qnA.getQContent();
 		Matcher matcher = ckImagePattern.matcher(content);
 		while(matcher.find()) {
 			String src = matcher.group();
@@ -68,17 +87,41 @@ public class QnARestService {
 			File file = new File(imageFolder, fileName);
 			if(file.exists()==true)
 				file.delete();
-		}*/
+		}
 		qdao.delete(qNo);
 	}
 	public void updateQnA(QnADto.DtoForUpdate dto) {
 		QnA qna = qdao.findById(dto.getQNo());
+		if(qna==null)
+			throw new JobFailException("문의를 찾을 수 없습니다");
+		dto.setMUsername(mdao.findMirumeByMnum(qna.getMNum()));
+		if(mdao.findMirumeByMnum(qna.getMNum()).equals(dto.getMUsername())==false)
+			throw new UserNotFoundException();
 		qna = mapper.map(dto, QnA.class);
 		qdao.update(qna);
 	}
 	
-	public int updateQnAcomment(QnAComment qnAComment){
-		return qndao.update(qnAComment);
+	public void updateQnAcomment(QnAComment qnAComment){
+		 qndao.update(qnAComment);
+	}
+	public String saveCkImage(MultipartFile upload) throws IOException {
+		Map<String,String> map = new HashMap<String, String>();
+		if(upload!=null) {
+			if(upload.getContentType().toLowerCase().startsWith("image/")){
+				String imageName = UUID.randomUUID().toString() + ".jpg";
+				File file = new File(imageFolder, imageName);
+				FileCopyUtils.copy(upload.getBytes(), file);
+				String fileUrl = imagePath + imageName;
+				// json 데이터로 등록
+				// {"uploaded" : 1, "fileName" : "test.jpg", "url" : "/img/test.jpg"}
+				// 이런 형태로 리턴이 나가야함.
+				map.put("uploaded", "1");
+				map.put("fileName", imageName);
+				map.put("url", fileUrl);
+                return objectmapper.writerWithDefaultPrettyPrinter().writeValueAsString(map);
+			}
+		}
+		return null;
 	}
 	
 }
